+++ db/migrate/20130326195342_add_total_steps_to_exports.rb
@@ -1,0 +1,7 @@
+class AddTotalStepsToExports < ActiveRecord::Migration
+  def change
+    add_column :exports, :steps_completed, :integer, null: false, default: 0
+    add_column :exports, :total_steps, :integer, null: false, default: 0
+    add_column :exports, :sheet_ids_count, :integer, null: false, default: 0
+  end
+end